+++ src/main/java/local/sgs/javatodos/Seeddata.java
@@ -1,9 +1,11 @@
 package local.sgs.javatodos;
 
 import local.sgs.javatodos.models.Role;
+import local.sgs.javatodos.models.Todo;
 import local.sgs.javatodos.models.User;
 import local.sgs.javatodos.models.UserRoles;
 import local.sgs.javatodos.repository.RoleRepository;
+import local.sgs.javatodos.repository.TodoRepository;
 import local.sgs.javatodos.repository.UserRepository;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
@@ -13,15 +15,15 @@ import java.util.ArrayList;
 
 @Transactional
 @Component
-public class SeedData implements CommandLineRunner {
+public class Seeddata implements CommandLineRunner {
     RoleRepository rolerepos;
     UserRepository userrepos;
-    ToDoRepository todorepos;
+    TodoRepository todorepos;
 
-    public SeedData(RoleRepository rolerepos, UserRepository userrepos, ToDoRepository todorepos) {
+    public Seeddata(RoleRepository rolerepos, UserRepository userrepos, TodoRepository todorepos) {
         this.rolerepos = rolerepos;
         this.userrepos = userrepos;
-        this.todorepos = todorepos
+        this.todorepos = todorepos;
     }
 
     @Override
@@ -44,12 +46,12 @@ public class SeedData implements CommandLineRunner {
         User u4 = new User("Jane", "password", users);
 
         // the date and time string should get coverted to a datetime Java data type. This is done in the constructor!
-        u4.getToDos().add(new ToDo("Finish java-orders-swagger", "2019-01-13 04:04:04", u4));
-        u4.getToDos().add(new ToDo("Feed the turtles", "2019-03-01 04:04:04", u4));
-        u4.getToDos().add(new ToDo("Complete the sprint challenge", "2019-02-22 04:04:04", u4));
+        u4.getTodos().add(new Todo("Finish java-orders-swagger", "2019-01-13 04:04:04", u4));
+        u4.getTodos().add(new Todo("Feed the turtles", "2019-03-01 04:04:04", u4));
+        u4.getTodos().add(new Todo("Complete the sprint challenge", "2019-02-22 04:04:04", u4));
 
-        u3.getToDos().add(new ToDo("Walk the dogs", "2019-01-17 04:04:04", u3));
-        u3.getToDos().add(new ToDo("provide feedback to my instructor", "2019-02-13 04:04:04", u3));
+        u3.getTodos().add(new Todo("Walk the dogs", "2019-01-17 04:04:04", u3));
+        u3.getTodos().add(new Todo("provide feedback to my instructor", "2019-02-13 04:04:04", u3));
 
         userrepos.save(u1);
         userrepos.save(u2);
