+++ src/main/java/local/sgs/javatodos/repository/TodoRepository.java
@@ -1,0 +1,8 @@
+package local.sgs.javatodos.repository;
+
+import local.sgs.javatodos.models.Todo;
+import org.springframework.data.repository.CrudRepository;
+
+public interface TodoRepository extends CrudRepository<Todo, Long>
+{
+}
